refac: share chat model
